apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic-settings==2.10.1
email-validator==2.3.0
httpx==0.28.1
python-jose[cryptography]==3.5.0
passlib==1.7.4
psycopg[binary]==3.2.10
